Add glossaries for #1886
- #1886 glossary loading added
- add all glossaries to the /glossary api
- created glossary migration
- load-data glossary loading
